Unify naming conventions, improve readme
